Move visit::Binding into separate file
